7 Basic Java Practice Questions Explained
7 basic java problems done in a simple straightforward manner. 
Commented to show understanding of concepts used, as well as provide notes for future repeated reading.
Everything works as required. Main focus was not writing the code itself, but to focus on being able to see Java in action. Assembled piece by piece and "commenting" on any new or difficult topics.